Change Database holder to allow for stub/testable database
I changed the database factory to allow for the creation
of the production database, a real database that we can test
and a stub database. I changed the holder to allow you
to manully put in your own database.
I changed the tests to put in a stub database into the holder.